[TASK] Move files in preparation for server separation
