+++ api/lib/errors.js
