+++ api/lib/logger.js
