+++ api/lib/schema-validator.js
@@ -2,7 +2,7 @@ var fs  = require('fs');
 var path = require('path');
 var jayschema = require('jayschema');
 
-var baseDir = path.join(__dirname, '/../schemas');
+var baseDir = path.join(__dirname, '../../schemas');
 
 module.exports = (function() {
   var validator = new jayschema();
