+++ api/orders.js
@@ -1,14 +1,14 @@
 var _                       = require('lodash');
 var Promise                 = require('bluebird');
 var ripple                  = require('ripple-lib');
-var remote                  = require('./../lib/remote.js');
+var remote                  = require('./lib/remote.js');
 var transactions            = require('./transactions');
-var SubmitTransactionHooks  = require('./../lib/submit_transaction_hooks.js');
-var respond                 = require('./../lib/response-handler.js');
-var utils                   = require('./../lib/utils');
-var errors                  = require('./../lib/errors.js');
-var TxToRestConverter       = require('./../lib/tx-to-rest-converter.js');
-var validator               = require('./../lib/schema-validator.js');
+var SubmitTransactionHooks  = require('./lib/submit_transaction_hooks.js');
+var respond                 = require('../server/response-handler.js');
+var utils                   = require('./lib/utils');
+var errors                  = require('./lib/errors.js');
+var TxToRestConverter       = require('./lib/tx-to-rest-converter.js');
+var validator               = require('./lib/schema-validator.js');
 var bignum                  = require('bignumber.js');
 
 const InvalidRequestError   = errors.InvalidRequestError;
