+++ api/payments.js
@@ -3,18 +3,18 @@ var async                   = require('async');
 var bignum                  = require('bignumber.js');
 var ripple                  = require('ripple-lib');
 var transactions            = require('./transactions');
-var validator               = require('./../lib/schema-validator');
-var remote                  = require('./../lib/remote.js');
-var serverLib               = require('./../lib/server-lib');
-var utils                   = require('./../lib/utils');
-var remote                  = require('./../lib/remote.js');
-var dbinterface             = require('./../lib/db-interface.js');
-var config                  = require('./../lib/config.js');
-var RestToTxConverter       = require('./../lib/rest-to-tx-converter.js');
-var TxToRestConverter       = require('./../lib/tx-to-rest-converter.js');
-var SubmitTransactionHooks  = require('./../lib/submit_transaction_hooks.js');
-var respond                 = require('./../lib/response-handler.js');
-var errors                  = require('./../lib/errors.js');
+var validator               = require('./lib/schema-validator');
+var remote                  = require('./lib/remote.js');
+var serverLib               = require('./lib/server-lib');
+var utils                   = require('./lib/utils');
+var remote                  = require('./lib/remote.js');
+var dbinterface             = require('./lib/db-interface.js');
+var config                  = require('./lib/config.js');
+var RestToTxConverter       = require('./lib/rest-to-tx-converter.js');
+var TxToRestConverter       = require('./lib/tx-to-rest-converter.js');
+var SubmitTransactionHooks  = require('./lib/submit_transaction_hooks.js');
+var respond                 = require('../server/response-handler.js');
+var errors                  = require('./lib/errors.js');
 
 var InvalidRequestError     = errors.InvalidRequestError;
 var NetworkError            = errors.NetworkError;
