+++ package.json
@@ -4,7 +4,7 @@
   "description": "A RESTful API for submitting payments and monitoring accounts on the Ripple network.",
   "scripts": {
     "test": "NODE_ENV=test ./node_modules/.bin/istanbul test ./node_modules/.bin/_mocha -- --ui tdd --reporter spec --timeout 10000 --slow 500 test/*-test.js test/unit/*-test.js",
-    "start": "node server.js",
+    "start": "node server/server.js",
     "coveralls": "cat ./coverage/lcov.info | ./node_modules/.bin/coveralls"
   },
   "repository": {
