+++ test/_payments-test.js
@@ -1,5 +1,5 @@
 var supertest     = require('supertest');
-var _app          = require('./../lib/express_app');
+var _app          = require('../server/express_app');
 var app           = supertest(_app);
 var assert        = require('assert');
 var ws            = require('ws');
